Allow the MQTT client ID to be set in the configuration file
This introduces a new MQTT_ClientID option to the config file
which allows you to set a separate MQTT Client ID value. This
is important as MQTT will only allow one client to connect
with each client ID, so if you have multiple instances of the
PiSomfy running, when each one connects without this change
it will disconnect any other instances from the MQTT broker.